+++ tensorflow_recommenders/experimental/optimizers/__init__.py
@@ -1,4 +1,4 @@
-# Copyright 2022 The TensorFlow Recommenders Authors.
+# Copyright 2026 The TensorFlow Recommenders Authors.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -14,4 +14,5 @@
 
 """Experimental Optimizers."""
 
+from tensorflow_recommenders.experimental.optimizers.clippy_adagrad import ClippyAdagrad
 from tensorflow_recommenders.experimental.optimizers.composite_optimizer import CompositeOptimizer