+++ tensorflow_recommenders/layers/feature_interaction/dot_interaction_test.py
@@ -1,4 +1,4 @@
-# Copyright 2022 The TensorFlow Recommenders Authors.
+# Copyright 2026 The TensorFlow Recommenders Authors.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -64,7 +64,7 @@ class DotInteractionTest(tf.test.TestCase):
                                      f13, f23, 0]]), output)
 
   def test_non_matching_dimensions(self):
-    with self.assertRaisesRegexp(ValueError, r"dimensions must be equal"):
+    with self.assertRaisesRegex(ValueError, r"dimensions must be equal"):
       feature1 = np.asarray([[0.1, 0.2, 0.3]]).astype(np.float32)
       feature2 = np.asarray([[2.0, -1.0, 1.0]]).astype(np.float32)
       feature3 = np.asarray([[0.0, 1.0]]).astype(np.float32)