+++ tools/build_scripts/pip_install.sh
@@ -7,23 +7,25 @@ set -e   # fail and exit on any command erroring
 : "${TF_VERSION:?}"
 : "${PY_VERSION:?}"
 
-# Import build functions.
-source ./tools/build_scripts/utils.sh
+# Use keras-2
+export TF_USE_LEGACY_KERAS=1
 
 which python3.10
 python3.10 --version
 
 # Install pip
 echo "Upgrading pip."
-pip install --upgrade pip
+python3.10 -m pip install --upgrade pip
 
 # Install TensorFlow.
 echo "Installing TensorFlow..."
-install_tf "$TF_VERSION"
+python3.10 -m pip install tensorflow
+python3.10 -m pip install -q tf-keras
+python3.10 -m pip install -q urllib3
 
 # Install TensorFlow Recommenders.
 echo "Installing TensorFlow Recommenders..."
-pip install -e .[docs]
+python3.10 -m pip install -e .[docs]
 
 # Test successful build.
 echo "Testing import..."